improve spot instances upgrades

+++ service/controller/azuremachinepool/handler/nodepool/create_terminate_old_workers.go
@@ -33,36 +33,50 @@ func (r *Resource) terminateOldWorkersTransition(ctx context.Context, obj interf
 		return currentState, microerror.Mask(err)
 	}
 
-	virtualMachineScaleSetsClient, err := r.ClientFactory.GetVirtualMachineScaleSetsClient(ctx, azureMachinePool.ObjectMeta)
-	if err != nil {
-		return currentState, microerror.Mask(err)
-	}
+	if len(oldInstances) > 0 {
+		r.Logger.Debugf(ctx, "There are still %d workers from the previous release running", len(oldInstances))
 
-	var ids compute.VirtualMachineScaleSetVMInstanceRequiredIDs
-	{
-		var strIds []string
-		for _, i := range oldInstances {
-			strIds = append(strIds, *i.InstanceID)
+		r.Logger.Debugf(ctx, "terminating %d old worker instances", len(oldInstances))
+
+		var ids compute.VirtualMachineScaleSetVMInstanceRequiredIDs
+		{
+			var strIds []string
+			for _, i := range oldInstances {
+				strIds = append(strIds, *i.InstanceID)
+			}
+
+			ids = compute.VirtualMachineScaleSetVMInstanceRequiredIDs{
+				InstanceIds: to.StringSlicePtr(strIds),
+			}
 		}
 
-		ids = compute.VirtualMachineScaleSetVMInstanceRequiredIDs{
-			InstanceIds: to.StringSlicePtr(strIds),
+		virtualMachineScaleSetsClient, err := r.ClientFactory.GetVirtualMachineScaleSetsClient(ctx, azureMachinePool.ObjectMeta)
+		if err != nil {
+			return currentState, microerror.Mask(err)
 		}
+
+		res, err := virtualMachineScaleSetsClient.DeleteInstances(ctx, key.ClusterID(&azureMachinePool), key.NodePoolVMSSName(&azureMachinePool), ids)
+		if err != nil {
+			return currentState, microerror.Mask(err)
+		}
+		_, err = virtualMachineScaleSetsClient.DeleteInstancesResponder(res.Response())
+		if err != nil {
+			return currentState, microerror.Mask(err)
+		}
+
+		r.Logger.Debugf(ctx, "terminated %d old worker instances", len(oldInstances))
+
+		return currentState, nil
 	}
 
-	r.Logger.Debugf(ctx, "filtered instance IDs for old instances")
-	r.Logger.Debugf(ctx, "terminating %d old worker instances", len(*ids.InstanceIds))
+	// All old nodes are terminated.
+	r.Logger.Debugf(ctx, "no old workers were found")
 
-	res, err := virtualMachineScaleSetsClient.DeleteInstances(ctx, key.ClusterID(&azureMachinePool), key.NodePoolVMSSName(&azureMachinePool), ids)
+	// Enable cluster autoscaler for this nodepool.
+	err = r.enableClusterAutoscaler(ctx, azureMachinePool)
 	if err != nil {
 		return DeploymentUninitialized, microerror.Mask(err)
 	}
-	_, err = virtualMachineScaleSetsClient.DeleteInstancesResponder(res.Response())
-	if err != nil {
-		return DeploymentUninitialized, microerror.Mask(err)
-	}
 
-	r.Logger.Debugf(ctx, "terminated %d old worker instances", len(*ids.InstanceIds))
-
-	return WaitForOldWorkersToBeGone, nil
+	return DeploymentUninitialized, nil
 }
